requirements.txt
torch>=2.1.0
torchvision>=0.16.0
opencv-python
numpy>=1.23.0,<2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
Pillow>=9.5.0
kaggle
